% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_response.R
\name{.check_response}
\alias{.check_response}
\title{.check_response}
\usage{
.check_response(resp)
}
\arguments{
\item{resp}{The httr response object.}
}
\value{
If the request is successful, returns \code{TRUE}, otherwise the
program is stopped with a message.
}
\description{
An internal function that checks the httr GET response.
}
\keyword{internal}
