% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_url.R
\name{.get_url}
\alias{.get_url}
\title{.get_url}
\usage{
.get_url(dataset, vintage)
}
\arguments{
\item{dataset}{A string that sets the name of the data set of interest (e.g. "acs/acs5").
This is a required parameter.}

\item{vintage}{An optional numeric that sets the vintage of interest.}
}
\value{
Returns a url string.
}
\description{
An internal function that creates the base url
address to be submitted as a http GET request by \code{.get_dt}.
It initializes with the dataset name and optionally the vintage.
}
\keyword{internal}
