% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_response.R
\name{.parse_response}
\alias{.parse_response}
\title{.parse_response}
\usage{
.parse_response(resp)
}
\arguments{
\item{resp}{The httr response object}
}
\value{
Converted json content to R objects.
}
\description{
An internal function checks for valid json content
from the Bureau's response. The content is converted to R objects.
}
\keyword{internal}
