% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_tiger_url.R
\name{.send_tiger_url}
\alias{.send_tiger_url}
\title{.send_tiger_url}
\usage{
.send_tiger_url(
  a_url = NULL,
  output_dir = NULL,
  crs_transform = NULL,
  sf_info = TRUE,
  do_progress = FALSE,
  caller = NULL
)
}
\arguments{
\item{a_url}{A required string that defines the url to be sent in the http GET request.}

\item{output_dir}{A required parameter that sets the full directory path for the download
shapefile destination. The files in this directory are removed before the download.}

\item{crs_transform}{A numeric which if non-NULL calls sf::st_transform()
to perform a crs transform of the geometries.}

\item{sf_info}{A logical which if TRUE displays info on the resulting simple feature object.}

\item{do_progress}{A logical which if TRUE displays a progress bar during the download.}

\item{caller}{A string that identifies the function making the call.}
}
\value{
A data frame object of class sf, data frame
}
\description{
An internal function for sending http requests to
the US Census Bureau TIGER api via httr::GET() function.
}
\author{
Rick Dean
}
\keyword{internal}
