% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_category_strings.R
\name{get_category_strings}
\alias{get_category_strings}
\title{get_category_strings}
\usage{
get_category_strings(
  name = NULL,
  get_names = FALSE,
  start_idx = NULL,
  end_idx = NULL
)
}
\arguments{
\item{name}{A string that is the name of the categorical variable of interest. This is a required parameter.}

\item{get_names}{An optional logical which if TRUE the function will return
a vector of categorical variable names recognized by the function.}

\item{start_idx}{An optional integer that sets the starting index of the variable's integer/label pairs.
If the value is NULL(the default), then the variable's entire integer/label pairs are returned.}

\item{end_idx}{An optional integer that sets the ending row index of the returned variable's integer/label pairs.}
}
\value{
A data.table with the variable's integer values and corresponding label strings.
}
\description{
For a Census Bureau categorical variable return a data.table with the
variable's integer values and corresponding label strings.
}
\details{
The function will attempt to locate a Bureau's categorical variable by name and
return a data.table with both the integer values and corresponding label strings. If
the variable is not located then NULL is returned.  Definitions for the variables were
obtained from
\href{https://www.census.gov/data/developers/data-sets/popest-popproj/popest/popest-vars.Vintage_2019.html}{Population Estimates Categorical Variables}
and other Bureau sources.
}
\author{
Rick Dean
}
