% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geography.R
\name{get_geography}
\alias{get_geography}
\title{get_geography}
\usage{
get_geography(dataset, vintage = NULL)
}
\arguments{
\item{dataset}{A required string that sets the acronym name of the data set of interest (e.g. "acs/acs5")}

\item{vintage}{An optional numeric that sets the year of interest.}
}
\value{
A data.table
}
\description{
Get the list of geography entities available (state, county, tract, etc) for a
a specific dataset.
}
\details{
Function produces a data.table of all the geography "name" and "geoLevelDisplay" variables
available for a specific dataset and optionally a vintage.
}
\author{
Rick Dean
}
