% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_idb_data.R
\name{get_idb_data}
\alias{get_idb_data}
\title{get_idb_data}
\usage{
get_idb_data(
  dataset = NULL,
  years = 2023,
  vars = NULL,
  group = FALSE,
  wide_to_long = FALSE,
  countries = NULL,
  ages = 1:100,
  sex = c(0, 1, 2),
  key = Sys.getenv("CENSUS_KEY")
)
}
\arguments{
\item{dataset}{A required string that selects the dataset of interest. Acceptable values are
"1year" or "5year".}

\item{years}{An optional numeric vector that sets the years of interest. The default is 2023.}

\item{vars}{An optional string vector of variable names to be acquired. The default is GEO_ID,NAME,GENC,POP
for either "1year" or "5year".}

\item{group}{A logical that if TRUE returns all the variables from the "5year" dataset. The default is FALSE.}

\item{wide_to_long}{The returned data.table is normally in a wide format with all the variables as columns.
If this logical parameter is TRUE then a long format is returned with variable names in one column (named "estimate") and
their respective values in another column (named "value").}

\item{countries}{An optional string vector of country/area abbreviations (GENC Standard Countries and Areas) of interest.}

\item{ages}{An optional numeric that selects specific ages(0 to 100 range) from the 1year dataset. Default is all ages.}

\item{sex}{An optional numeric vector to select the sex of interest from the 1year dataset. Acceptable values
0 - Both, 1 - Male, 2 - Female. Default is all three values for sex.}

\item{key}{A string that sets the access key. All Census Bureau API requests require an access key.
Sign-up for a key is free and can be obtained \href{https://api.census.gov/data/key_signup.html}{here}.
The function will check for a global setting of the key via \code{Sys.getenv("CENSUS_KEY")}.
Run \code{usethis::edit_r_environ()} and edit your .Renviron file with the line: CENSUS_KEY=your key
to create the global association. This is a required parameter.}
}
\value{
A data.table
}
\description{
Get data from the Census Bureau's Time Series International Database
for a specfic dataset (\emph{Single Year of Age and Sex} or \emph{5-Year Age Groups and Sex}),
year, variables, and country.
}
\details{
See \href{https://api.census.gov/data/timeseries/idb/1year/variables.html}{1year} and
\href{https://api.census.gov/data/timeseries/idb/5year/variables.html}{5year} for variable
names.

See Census Bureau's publicly available
\href{https://www.census.gov/data/developers/data-sets/international-database.html}{International Database} for
dataset descriptions.
}
\author{
Rick Dean
}
