% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_multi_vintage_data.R
\name{get_multi_vintage_data}
\alias{get_multi_vintage_data}
\title{get_multi_vintage_data}
\usage{
get_multi_vintage_data(
  dataset = NULL,
  vintage_v = NULL,
  vars = NULL,
  region = NULL,
  regionin = NULL,
  key = Sys.getenv("CENSUS_KEY")
)
}
\arguments{
\item{dataset}{A string that sets the name of the data set of interest (e.g. "acs/acs5"). This is a required parameter.}

\item{vintage_v}{A numeric vector that sets the vintages of interest. This is a required parameter.}

\item{vars}{A string vector of variable acronym names to be acquired (e.g. "B15002_015").
See RcensusPkg::get_variable_names() for obtaining acronym names.  This is a required parameter.}

\item{region}{A string that specifies the geography of the request. See \code{Rcensus::get_geography()} for
assistance in obtaining these values.}

\item{regionin}{A string that sets a qualifier for \code{region}.}

\item{key}{A string that sets the access key. All Census Bureau API requests require an access key.
Sign-up for a key is free and can be obtained \href{https://api.census.gov/data/key_signup.html}{here}.
The function will check for a global setting of the key via \code{Sys.getenv("CENSUS_KEY")}.
This is a required parameter.}
}
\value{
A data.table
}
\description{
Get Census Bureau data for a specific dataset, variables, and region
in the form of a dataframe for multiple vintages.
}
\details{
Function produces a data.table/dataframe of
Census Bureau data for multiple vintages.  The function requires an access key
issued from the Bureau.
}
\author{
Rick Dean
}
