% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variable_names.R
\name{get_variable_names}
\alias{get_variable_names}
\title{get_variable_names}
\usage{
get_variable_names(
  dataset = NULL,
  vintage = NULL,
  vars = NULL,
  group = NULL,
  filter_group_est = FALSE,
  filter_name_str = NULL,
  filter_label_str = NULL,
  filter_concept_str = NULL,
  ignore_case = TRUE,
  fixed = FALSE
)
}
\arguments{
\item{dataset}{A required string that sets the name of the dataset of interest (e.g. "acs/acs5").
See \code{Rcensus::get_dataset_names()} for available dataset names.}

\item{vintage}{An required numeric that sets the year of interest.}

\item{vars}{An optional vector of variable names whose descriptions are of interest.}

\item{group}{An optional string that sets the group name associated with a set of variables.
See \code{Rcensus::get_groups()} for available group names under a specific dataset and vintage.}

\item{filter_group_est}{A logical which if TRUE will filter the variable names from \code{group} and return
only estimate related variable names. The default is FALSE.}

\item{filter_name_str}{A character string by which to filter the resultant data.table's "name" column.}

\item{filter_label_str}{A character string by which to filter the resultant data.table's "label" column.}

\item{filter_concept_str}{A character string by which to filter the resultant data.table's "concept" column.}

\item{ignore_case}{A logical which if FALSE will not ignore case in filtering the "name", "label", "concept" column.}

\item{fixed}{A logical which if TRUE, then the above filter strings are used 'as is' in matching.}
}
\value{
A data.table
}
\description{
Get Census Bureau variable acronym names and their label descriptions.
}
\details{
Function produces a data.table of variable acronym names and their
descriptions from the Census Bureau's publicly available
\href{https://www.census.gov/data/developers/data-sets.html}{datasets}.
The function returns 4 columns:
\describe{
\item{name}{the name of the parameter}
\item{label}{the Burau's description of the parameter}
\item{required}{a boolean indicating if the parameter is required}
\item{predicateType}{a string indicating the variable primitive type}
}

Note that a variable with a "required" character value "true" must be included
in your data requests (i.e. \code{RcensusPkg::get_vintage_data()}) or
the API will return an error.
}
\author{
Rick Dean
}
