% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vintage_data.R
\name{get_vintage_data}
\alias{get_vintage_data}
\title{get_vintage_data}
\usage{
get_vintage_data(
  dataset = NULL,
  vintage = NULL,
  vars = NULL,
  NAME_GEOID = TRUE,
  predicates = NULL,
  group = NULL,
  region = NULL,
  regionin = NULL,
  na_cols = NULL,
  key = Sys.getenv("CENSUS_KEY")
)
}
\arguments{
\item{dataset}{A string that sets the name of the data set of interest (e.g. "acs/acs5").
See Census Bureau's publicly available
\href{https://www.census.gov/data/developers/data-sets.html}{datasets} for
descriptions. Descriptions/vintages for datasets can also be found by running
\code{Rcensus::get_dataset_names()}. This is a required parameter.}

\item{vintage}{An optional numeric that sets the vintage of interest. Available vintages
for a specific dataset can be found by running \code{Rcensus::get_dataset_names()}.}

\item{vars}{A string vector of variable names to be acquired.
Available variable names can be determined by running \code{Rcensus::get_variable_names()}.}

\item{NAME_GEOID}{A logical which if TRUE will add "NAME" and "GEO_ID" variables to \code{vars} string vector.
The default is TRUE.}

\item{predicates}{An optional vector of strings that adds data filtering.
See \href{https://www.census.gov/data/developers/guidance.html}{Census Data API User Guide} for
forming predicates and filtering or limiting variables. As noted in the guide each predicate must
start with an ampersand sign.}

\item{group}{An optional string that names an entire group of similar variables to be retrieved.
For example the group value "B01001" would return values of all variables related to
"SEX BY AGE". To find available groups submit a dataset and vintage to \code{RcensusPkg::get_groups}.}

\item{region}{An optional string that specifies the geography of the request.
See \href{https://www.census.gov/data/developers/guidance.html}{Census Data API User Guide} for
specifying this and the \code{regionin} parameter. Not all regions such as counties,
blocks, or tracts are available for a specific dataset and vintage. Use
\code{Rcensus::get_geography()} to check on both \code{region} and \code{regionin}.}

\item{regionin}{A string that sets a qualifier for \code{region}.}

\item{na_cols}{If TRUE will remove all rows with missing values. If a
vector of column names/integers will check only those columns for missing values.}

\item{key}{A string that sets the access key. All Census Bureau API requests require an access key.
Sign-up for a key is free and can be obtained \href{https://api.census.gov/data/key_signup.html}{here}.
The function will check for a global setting of the key via \code{Sys.getenv("CENSUS_KEY")}.
Run \code{usethis::edit_r_environ()} and edit your .Renviron file with the line: CENSUS_KEY=your key
to create the global association. This is a required parameter.}
}
\value{
A data.table
}
\description{
Get Census Bureau data for a specific dataset, variables, and region
in the form of a data.table.
}
\details{
Function produces a data.table with selected Census Bureau variables as columns. The function requires an access key
issued from the Bureau. Variables of interest can be specified individually or by group/table name.
Predicate phrases can be specified for filtering the results.
}
\author{
Rick Dean
}
