% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_it.R
\name{join_it}
\alias{join_it}
\title{join_it}
\usage{
join_it(
  df_1 = NULL,
  df_2 = NULL,
  key_1 = NULL,
  key_2 = NULL,
  negate = FALSE,
  match = FALSE,
  return_sf = FALSE,
  na_rm = TRUE
)
}
\arguments{
\item{df_1}{The first dataframe to be joined.}

\item{df_2}{The second dataframe to be joined with \code{df_1}. All rows in
this dataframe will be present in the resultant dataframe.}

\item{key_1}{A string that names the column from \code{df_1} that is common to \code{df_2}.}

\item{key_2}{A string that names the column from \code{df_2} that is common to \code{df_1}.}

\item{negate}{An optional logical which if TRUE will return a dataframe
that has rows in \code{df_1} but not in \code{df_2}.}

\item{match}{An optional logical which if TRUE will return a dataframe
that has rows where only both \code{df_1} and \code{df_2} have matches.}

\item{return_sf}{An optional logical which if TRUE will convert the resultant
data.table to a simple feature if it has a geometries column.}

\item{na_rm}{An optional logical which if TRUE (the default) then remove rows
with NA values.}
}
\value{
A data.table or simple feature object if \code{return_sf} is TRUE.
}
\description{
Outer join two dataframes that have a common column variable.
}
\details{
Function uses fast data.table techniques to join two data.tables
by their common key values.  Examples might include using the "GEOID" variable
as a key to join data from RcensusPkg::get_vintage_data() with a
simple feature with its geometries for counties, states, countries for example
from RcensusPkg::tiger_\emph{\emph{sf(). The resulting dataframe could then display the
geometries (with RspatialPkg::get_geom_sf()) with an aesthetic mapping
(e.g. fill/color/size) with a joined data column. Joining could also take place
between two simple features (created by RcensusPkg::tiger}}_sf()) or between
two dataframes created by (created by RcensusPkg::get_vintage_data()).

The important thing to remember is that all the rows in \code{df_2} will
be present in the resultant data.table.
}
\author{
Rick Dean
}
