% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_to_wide.R
\name{long_to_wide}
\alias{long_to_wide}
\title{long_to_wide}
\usage{
long_to_wide(
  dt = NULL,
  id_v = c("NAME", "GEOID"),
  parameter_col = NULL,
  value_col = NULL
)
}
\arguments{
\item{dt}{The data.table with a long column format. This parameter is required.}

\item{id_v}{A vector of column names from \code{dt} that act as identifiers and
are not part of the widened column. This parameter is required.}

\item{parameter_col}{A column name from \code{dt} whose unique values will become
column names for the new expanded data.table.}

\item{value_col}{A column name or vector of column names from \code{dt} whose values will fall under
the new expanded data.table. This parameter is required.}
}
\value{
A reshaped data.table in the "wide" format.
}
\description{
Reshape a data.table from a "long" format to a "wide" format.
}
\details{
Function calls data.table::dcast() to reshape a long single column and
its values to multilpe colums.
}
\author{
Rick Dean
}
