% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_us_data.R
\name{plot_us_data}
\alias{plot_us_data}
\title{plot_us_data}
\usage{
plot_us_data(
  df = NULL,
  states_col = NULL,
  value_col = NULL,
  title = NULL,
  text_col = NULL,
  text_size = 3,
  text_color = "black",
  text_fontface = "plain",
  output_dir = tempdir(check = T),
  vintage = 2020,
  general = FALSE,
  resol = "500k",
  hide_x_tics = TRUE,
  hide_y_tics = TRUE,
  show_legend = TRUE,
  legend_pos = "right",
  check_na = FALSE,
  scale_breaks = NULL,
  scale_values = NULL,
  scale_limits = NULL,
  scale_labels = NULL,
  scale_palette = 1,
  scale_direction = -1,
  scale_na_value = "gray50",
  sf_color = "black",
  sf_fill = "gray",
  sf_linewidth = 0.1,
  sf_alpha = 1,
  display_plot = TRUE
)
}
\arguments{
\item{df}{The data frame with a column of full state names and a second variable column of their respective values.
The column name for the states must be "NAME".}

\item{states_col}{A string that sets the column name from \code{df} containing the state names of interest. These are
full state names, either capitalized or lower case. This is a required parameter.}

\item{value_col}{A string that sets the column name from \code{df} where values(discrete or continuous) are defined.
If the column has discrete values then it must be a factor. This is a required parameter.}

\item{title}{A string that sets the plot title.}

\item{text_col}{An optional string that sets the column name from \code{df} for labelling each state polygon.}

\item{text_size}{A numeric value that sets the size of labeled state text.}

\item{text_color}{A string that sets the color of labeled state text.}

\item{text_fontface}{A string that sets the fontface of labeled state text.
Acceptable values: "plain", "bold", "italic", "bold.italic". The default is "plain".}

\item{output_dir}{A full directory path where the shapefile and its associated files will be downloaded.
The default is the directory defined by the value returned by \code{tempdir()}.}

\item{vintage}{A numeric that sets the vintage of interest. The default is 2020.}

\item{general}{A logical which if TRUE will download a less detailed, more generalized version of the state geometries.}

\item{resol}{If \code{general} is TRUE, then the resolution to return. Acceptable values are strings
"500k", "5m", "20m".}

\item{hide_x_tics}{A logical that controls the appearance of the x axis tics.}

\item{hide_y_tics}{A logical that controls the appearance of the y axis tics.}

\item{show_legend}{A logical that controls the appearance of the legend.}

\item{legend_pos}{A string that sets the legend position. Acceptable values are "right",
"top", "bottom".}

\item{check_na}{A logical which if TRUE will remove rows that have missing values for any of the columns.
The default is to not check the columns for NA values.}

\item{scale_breaks}{A string/numeric vector that defines the scale breaks. This is a required parameter.}

\item{scale_values}{A string/numeric vector that defines the possible values. For factor values, this is required
and is a vector string of colors.}

\item{scale_limits}{A string/numeric vector that defines the scale limits.}

\item{scale_labels}{An optional string vector that defines the scale labels. Vector must be the same length
as \code{scale_breaks}.}

\item{scale_palette}{If a string, will use that named palette for continuous values. If a number, will index into a colorbrewer type list of
\code{sequential, diverging, qualitative}. See ggplot2::scale_fill_distiller() for more information.}

\item{scale_direction}{Sets the order of colors in the scale. If -1, the order is reversed.}

\item{scale_na_value}{A string that sets the color for missing values.}

\item{sf_color}{A string that sets the polygon border line color.}

\item{sf_fill}{A string that sets the polygon area fill color.}

\item{sf_linewidth}{A numeric that sets the border line thickness.}

\item{sf_alpha}{A numeric that sets the alpha level attribute of the polygon fill.}

\item{display_plot}{A logical that if TRUE will display the plot. The default is TRUE.
If FALSE, then a ggplot2 object is returned}
}
\value{
A list of ggplot2 objects if \code{display_plot} is FALSE. Included in the
list is the plot of all the states ("us_states") along with the original ggplot2
geom_sf plots of the lower 48 ("lower_48"), Alaska ("alaska"), Hawaii ("hawaii") and Puerto Rico ("puerto_rico").
}
\description{
This function produces a ggplot2 based choropleth map of a discrete/continuous variable across all/selected US states
including viewable placement of Alaska, Hawaii, and Puerto Rico.  The function accepts a data frame with a column of
state names and a column with their respective values.  The function offers several options for
control of state geographies and variable legend.
}
\details{
Because scaling is done manually, "scale_breaks" is a required parameter. For discrete values this must
must be a character vector. For continuous values this must be a numeric vector.

This function depends extensively on \code{RcensusPkg::tiger_states_sf()} for obtaining state geometries, so many of
that function's parameters are repeated in this function. Also \code{RspatialPkg::get_geom_sf()} is called upon for
displaying the shapefile geometries.
}
\author{
Rick Dean
}
