% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_area_water.R
\name{remove_area_water}
\alias{remove_area_water}
\title{remove_area_water}
\usage{
remove_area_water(x, vintage = 2020, output_dir = tempdir(check = T))
}
\arguments{
\item{x}{An sf shapefile object with possible areas of water downloaded from the Census Bureau}

\item{vintage}{An integer that specifies the year of the shapefile. The default is 2020.}

\item{output_dir}{A full directory path where supportive shapefiles and their associated files will be downloaded.
The default is the directory defined by the value returned by \code{tempdir()}.}
}
\value{
An sf shapefile object with the water area geometries removed.
}
\description{
Function removes water area geometries from downloaded Census Bureau shapefiles
}
\author{
Rick Dean
}
