% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiger_counties_sf.R
\name{tiger_counties_sf}
\alias{tiger_counties_sf}
\title{tiger_counties_sf}
\usage{
tiger_counties_sf(
  output_dir = NULL,
  vintage = 2020,
  general = FALSE,
  resol = "500k",
  crs_transform = NULL,
  sf_info = TRUE,
  do_progress = FALSE,
  shapefile = NULL,
  datafile = NULL,
  datafile_key = NULL,
  sf_key = "COUNTYFP",
  express = NULL,
  check_na = FALSE
)
}
\arguments{
\item{output_dir}{A full directory path where the shapefile will be downloaded. This is a required parameter. The function will stop
if this directory does not exist.  Be aware that all files in this directory are removed before downloading.}

\item{vintage}{A numeric that sets the vintage of interest. The default is 2020.}

\item{general}{A logical which if TRUE will download a less detailed, more generalized version of the county geometries.}

\item{resol}{If \code{general} is TRUE, then the resolution to return. Acceptable values are strings
"500k", "5m", "20m".}

\item{crs_transform}{A numeric or string that if non-NULL transforms the geometries to this coordinate reference system. See
\href{sf::st_transform()}{https://cran.r-project.org/web/packages/sf/sf.pdf} for acceptable values.}

\item{sf_info}{A logical which if TRUE displays info on the resulting simple feature object.}

\item{do_progress}{A logical which if TRUE displays a progress bar during the download.}

\item{shapefile}{A full file path to a shapefile folder with its unzipped files to be processed instead of downloading.}

\item{datafile}{A dataframe containing data that should be joined with this function's resultant simple feature object.}

\item{datafile_key}{The column name from \code{datafile} dataframe used to key with the \code{sf_key} column of the resultant simple feature dataframe.}

\item{sf_key}{The column from the resultant dataframe used to key with the \code{datafile} dataframe.}

\item{express}{A logical expression object used to filter the resultant simple feature dataframe.
For example, one of the columns of the resultant simple feature dataframe is "STATEFP".
If you wanted to return just the geometries for Florida (which has a fips code of "12"),
then you assign \code{express} equal to: expression(STATEFP == "12"). The expression will be
evaluated and only the geometries for Florida will be returned.}

\item{check_na}{A logical which if TRUE will remove rows that have missing values for any of the columns.
The default is to not check the columns for NA values.}
}
\value{
A data frame object of class sf, data frame
}
\description{
This function performs three tasks:
\enumerate{
\item Download to a temporary directory a zip file from the TIGER/Line Shapefiles database.
\item Unzip the zip file and locate the shape file of interest.
\item Read and convert the shape file to a simple feature object.
}
}
\details{
Returns simple feature (sf) of the entire US county boundary related geometric polygons,
provided by the US Census Bureau's TIGER/Line Shapefiles database. See
\href{https://r-spatial.github.io/sf/articles/sf1.html}{Simple Features for R}
for more information on simple features. Along with the geometries, additional county related
variables are provided.  See \href{https://www2.census.gov/geo/pdfs/maps-data/data/tiger/tgrshp2022/TGRSHP2022_TechDoc_F-R.pdf}{Appendix I-3. Record Layout: County Subdivision State-based Shapefile)}
for a description of county related variables of the sf file. For further information on the Census Bureau's shape files see
\href{https://www2.census.gov/geo/pdfs/maps-data/data/tiger/tgrshp2021/TGRSHP2021_TechDoc_Ch3.pdf}{About the 2021 TIGER/Line Shapefiles}.
From \href{https://www2.census.gov/geo/pdfs/maps-data/data/tiger/tgrshp2022/TGRSHP2022_TechDoc_Ch4.pdf}{Chapter 4.7 Counties and Equivalent Enties} --
"Counties and equivalent entities are primary legal divisions of states. In most states, these entities are
termed 'counties.'"

A more generalized, recognizable version of the county geometries that has less download size is also available.  For more information on cartographic boundary files see
\href{https://www.census.gov/programs-surveys/geography/technical-documentation/naming-convention/cartographic-boundary-file.html}{Cartographic Boundary File Description}.
These files are available for vintages greater than 2013 with resolution 1:500k, 1:5m, 1:20m meters.

Some earlier vintages may have NA for the crs so you may need to specify the \code{crs_transform} to 3426.  Also
you may be interested in using a state level crs. See \href{https://epsg.io/}{epsg.io} to search worldwide for crs.

The function returns the simple feature object which can easily be mapped (see \href{https://github.com/deandevl/RspatialPkg}{RspatialPkg::get_geom_sf()}) or
joined with US Census Bureau demographic data via the GEOID value.
}
\author{
Rick Dean
}
