% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiger_landmarks_sf.R
\name{tiger_landmarks_sf}
\alias{tiger_landmarks_sf}
\title{tiger_landmarks_sf}
\usage{
tiger_landmarks_sf(
  state = NULL,
  output_dir = NULL,
  vintage = 2020,
  entity = "point",
  crs_transform = NULL,
  sf_info = TRUE,
  do_progress = FALSE,
  shapefile = NULL,
  express = NULL,
  check_na = FALSE
)
}
\arguments{
\item{state}{The two-digit FIPS code for the state of interest. This is required parameter.
See \href{https://cran.r-project.org/web/packages/usmap/usmap.pdf}{usmap::fips function} for finding FIPS codes.}

\item{output_dir}{A full directory path where the shapefile will be downloaded. This is a required parameter. The function will stop
if this directory does not exist.  Be aware that all files in this directory are removed before downloading.}

\item{vintage}{A numeric that sets the vintage of interest. The default is 2020.
The value should be greater than 2010.}

\item{entity}{A character string that sets the specific landmark entity of interest. The
acceptable values are "area" or "point". The default is "point".}

\item{crs_transform}{A numeric or string that if non-NULL transforms the geometries to this coordinate reference system. See
\href{sf::st_transform()}{https://cran.r-project.org/web/packages/sf/sf.pdf} for acceptable values.}

\item{sf_info}{A logical which if TRUE displays info on the resulting simple feature object.}

\item{do_progress}{A logical which if TRUE displays a progress bar during the download.}

\item{shapefile}{A full file path to a shapefile folder with its unzipped files to be processed instead of downloading.}

\item{express}{A logical expression object used to filter the resultant simple feature dataframe.
For example, one of the columns of the resultant simple feature dataframe is "STATEFP".
If you wanted to return just the geometries for Nelson County, Kentucky (which has a fips code of "179"),
then you assign \code{express} equal to: expression(COUNTYFP == "179"). The expression will be
evaluated and only the landmark geometries for Nelson County will be returned.}

\item{check_na}{A logical which if TRUE will remove rows that have missing values for any of the columns.
The default is to not check the columns for NA values.

\strong{Note: Vintage must be greater than 2010.}}
}
\value{
A data frame object of class sf, data frame
}
\description{
This function performs three tasks:
\enumerate{
\item Download to a temporary directory a zip file from the TIGER/Line Shapefiles database.
\item Unzip the zip file and locate the shape file of interest.
\item Read and convert the shape file to a simple feature object.
}
}
\details{
Returns simple feature (sf) landmark related geometric polygons provided
by the US Census Bureau's TIGER/Line Shapefiles database. See
\href{https://r-spatial.github.io/sf/articles/sf1.html}{Simple Features for R}
for more information on simple features. Along with the geometries, additional landmark related
variables are provided.  See
\href{https://www2.census.gov/geo/pdfs/maps-data/data/tiger/tgrshp2022/TGRSHP2022_TechDoc_F-R.pdf}{Appendix K-1.K-2, Record Layout: Area/Point Landmark Shapefile}
for a description of landmark related area and point variables of the sf file.
For further information on the Census Bureau's shape files see
\href{https://www2.census.gov/geo/pdfs/maps-data/data/tiger/tgrshp2021/TGRSHP2021_TechDoc_Ch3.pdf}{About the 2021 TIGER/Line Shapefiles}.
From \href{https://www2.census.gov/geo/pdfs/maps-data/data/tiger/tgrshp2022/TGRSHP2022_TechDoc_Ch4.pdf}{Chapter 4.11 Landmark (Area and Point} --
"The Census Bureau includes landmarks in the MAF/TIGER System to locate special features and help
enumerators during field operations. Some of the more common landmark types include area landmarks
(e.g., airports, cemeteries, parks, and educational facilities) and point landmarks (e.g., schools and
churches)"

The function returns the simple feature object which can easily be mapped (see \href{https://github.com/deandevl/RspatialPkg}{RspatialPkg::get_geom_sf()}) or
joined with US Census Bureau demographic data via the GEOID value.

Some earlier vintages may have NA for the crs so you may need to specify the \code{crs_transform} to 3426.  Also
you may be interested in using a state level crs. See \href{https://epsg.io/}{epsg.io} to search worldwide for crs.
}
\author{
Rick Dean
}
