% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiger_states_sf.R
\name{tiger_states_sf}
\alias{tiger_states_sf}
\title{tiger_states_sf}
\usage{
tiger_states_sf(
  output_dir = NULL,
  vintage = 2020,
  general = FALSE,
  resol = "500k",
  crs_transform = NULL,
  sf_info = TRUE,
  do_progress = FALSE,
  shapefile = NULL,
  datafile = NULL,
  datafile_key = NULL,
  sf_key = "GEOID",
  express = NULL,
  check_na = FALSE
)
}
\arguments{
\item{output_dir}{A full directory path where the shapefile will be downloaded. This is a required parameter. The function will stop
if this directory does not exist.  Be aware that all files in this directory are removed before downloading.}

\item{vintage}{A numeric that sets the vintage of interest. The default is 2020.}

\item{general}{A logical which if TRUE will download a less detailed, more gerneralized version of the state geometries.}

\item{resol}{If \code{general} is TRUE, then the resolution to return. Acceptable values are strings
"500k", "5m", "20m".}

\item{crs_transform}{A numeric or string that if non-NULL transforms the geometries to this coordinate reference system. See
\href{sf::st_transform()}{https://cran.r-project.org/web/packages/sf/sf.pdf} for acceptable values.}

\item{sf_info}{A logical which if TRUE displays info on the resulting simple feature object.}

\item{do_progress}{A logical which if TRUE displays a progress bar during the download.}

\item{shapefile}{A full file path to a shapefile folder with its unzipped files to be processed instead of downloading.}

\item{datafile}{A dataframe containing data that should be joined with this function's resultant simple feature object.}

\item{datafile_key}{The column name from \code{datafile} used to key with the "tract" column of the resultant simple feature object.}

\item{sf_key}{The column from the resultant dataframe used to key with the \code{datafile} dataframe.}

\item{express}{A logical expression object used to filter the resultant simple feature dataframe.
For example, one of the columns of the resultant simple feature dataframe is "STATEFP".
If you wanted to return just the geometries for Florida (which has a fips code of "12"),
then you assign \code{express} equal to: expression(STATEFP == "12"). The expression will be
evaluated and only the geometries for Florida will be returned.}

\item{check_na}{A logical which if TRUE will remove rows that have missing values for any of the columns.
The default is to not check the columns for NA values.}
}
\value{
A data frame object of class sf, data frame
}
\description{
This function performs three tasks:
\enumerate{
\item Download to a temporary directory a zip file from the TIGER/Line Shapefiles database.
\item Unzip the zip file and locate the shape file of interest.
\item Read and convert the shape file to a simple feature object.
}
}
\details{
Returns simple feature (sf) of state boundary related geometric polygons,
provided by the US Census Bureau's TIGER/Line Shapefiles database. See
\href{https://r-spatial.github.io/sf/articles/sf1.html}{Simple Features for R}
for more information on simple features. Along with the geometries, additional state related
variables are provided.  See \href{https://www2.census.gov/geo/pdfs/maps-data/data/tiger/tgrshp2022/TGRSHP2022_TechDoc_F-R.pdf}{Appendix P-5. Record Layout: State and Equivalent Entity National Shapefile)}
for a description of state related variables of the sf file. For further information on the Census Bureau's shape files see
\href{https://www2.census.gov/geo/pdfs/maps-data/data/tiger/tgrshp2021/TGRSHP2021_TechDoc_Ch3.pdf}{About the 2021 TIGER/Line Shapefiles}.
From \href{https://www2.census.gov/geo/pdfs/maps-data/data/tiger/tgrshp2022/TGRSHP2022_TechDoc_Ch4.pdf}{Chapter 4.17 States and State Equivalent Entities} --
"States and equivalent entities are the primary governmental divisions of the United States. In addition to
the fifty states, the Census Bureau treats the District of Columbia, Puerto Rico, and the Island Areas
(American Samoa, the Commonwealth of the Northern Mariana Islands, Guam, and the U.S. Virgin
Islands) as statistical equivalents of states for the purpose of data presentation."

A more generalized, recognizable version of the state geometries that has less download size is also available.  For more information on cartographic boundary files see
\href{https://www.census.gov/programs-surveys/geography/technical-documentation/naming-convention/cartographic-boundary-file.html}{Cartographic Boundary File Description}.
These files are available for vintages greater than 2013 with resolution 1:500k, 1:5m, 1:20m meters.

The function returns the simple feature object which can easily be mapped (see \href{https://github.com/deandevl/RspatialPkg}{RspatialPkg::get_geom_sf()}) or
joined with US Census Bureau demographic data. To help incorporate data files, this function
has a \code{datafile} parameter which will be joined with the resultant simple feature object. The only
requirement is that a common "key" for joining exist between the data dataframe and the simple feature dataframe.

Some earlier vintages may have NA for the crs so you may need to specify the \code{crs_transform} to 3426.  Also
you may be interested in using a state level crs. See \href{https://epsg.io/}{epsg.io} to search worldwide for crs.
}
\author{
Rick Dean
}
