% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vote2020}
\alias{vote2020}
\title{vote2020 data}
\format{
Data frame with columns:
\describe{
\item{State}{character string full state name}
\item{Party}{character string with values "D" or "R"}
}
}
\usage{
vote2020
}
\description{
Dataframe shows national presidential voting results for 2020 by state.
Letters "D" and "R" under the Party column indicate respectively Democrate or Republican
win for a state.
}
\keyword{datasets}
