% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wide_to_long.R
\name{wide_to_long}
\alias{wide_to_long}
\title{wide_to_long}
\usage{
wide_to_long(
  dt = NULL,
  id_v = c("NAME", "GEOID"),
  measure_v = NULL,
  do_est = FALSE,
  do_moe = FALSE,
  do_est_moe = FALSE,
  variable_name = NULL,
  value_name = NULL
)
}
\arguments{
\item{dt}{The data.table with a wide collection of column variables. This parameter is required.}

\item{id_v}{A vector of column names from \code{dt} that are not to be
consolidated and act as identifier columns for the new long form. This parameter is required.}

\item{measure_v}{An optional vector of measure column names from \code{dt} that are to be consolidated.}

\item{do_est}{An optional logical which if TRUE will locate "estimate" columns in \code{dt} and
consolidate their values to the "long" format.}

\item{do_moe}{An optional logical which if TRUE will locate "margin of error" columns in \code{dt} and
consolidate their values to the "long" format.}

\item{do_est_moe}{An optional logical which if TRUE will locate both "estimate" and "margin of error" columns
in \code{dt} and consolidate their values to the "long" format.}

\item{variable_name}{An optional string that sets column name for the consolidated column names.}

\item{value_name}{An optional string that sets the column name for the consolidated values.}
}
\value{
A reshaped data.table in the "long" format.
}
\description{
Reshape a data.table from a "wide" format to a "long" format.
}
\details{
Function calls data.table::melt() to reshape a wide collection of data.table
columns and their values to a single long column of values.
}
\author{
Rick Dean
}
