% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_tiger_url.R
\name{.send_tiger_url}
\alias{.send_tiger_url}
\title{.send_tiger_url}
\usage{
.send_tiger_url(
  a_url = NULL,
  output_dir = NULL,
  delete_files = TRUE,
  set_crs = NULL,
  transform_crs = NULL,
  sf_info = FALSE,
  do_progress = FALSE,
  caller = NULL
)
}
\arguments{
\item{a_url}{A required string that defines the url to be sent in the http GET request.}

\item{output_dir}{A full directory path where the shapefile and its associated files will be downloaded.}

\item{delete_files}{A logical which if \code{TRUE} will delete the shapefile and associated files in 'output_dir'.
The default is \code{TRUE}.}

\item{set_crs}{A numeric or character string which if non-NULL calls \code{sf::st_crs()} to set the crs of the geometries.}

\item{transform_crs}{A numeric or character string which if non-NULL calls \code{sf::st_transform()}
to perform a crs transform of the geometries.}

\item{sf_info}{A logical which if \code{TRUE} displays info on the resulting simple feature object.}

\item{do_progress}{A logical which if \code{TRUE} displays a progress bar during the download.}

\item{caller}{A string that identifies the function making the call.}
}
\value{
A data frame object of class sf, data frame
}
\description{
An internal function for sending http requests to
the US Census Bureau TIGER api via httr::GET() function.
}
\keyword{internal}
