% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_us_data.R
\name{plot_us_data}
\alias{plot_us_data}
\title{plot_us_data}
\usage{
plot_us_data(
  df = NULL,
  states_col = NULL,
  value_col = NULL,
  title = NULL,
  title_fontsz = 14,
  text_col = NULL,
  text_size = 3,
  text_color = "black",
  text_fontface = "plain",
  output_dir = tempdir(check = T),
  vintage = 2020,
  general = FALSE,
  resol = "500k",
  show_legend = TRUE,
  legend_pos = "right",
  na_rm = FALSE,
  scale_breaks = NULL,
  scale_values = NULL,
  scale_limits = NULL,
  scale_labels = waiver(),
  scale_colors = heat.colors(8),
  scale_na_value = "gray50",
  own_scale = FALSE,
  sf_color = "black",
  sf_fill = "gray",
  sf_linewidth = 0.1,
  sf_alpha = 1,
  display_plot = TRUE
)
}
\arguments{
\item{df}{The data frame with a column of full state names and a second variable column of their respective values.
The column name for the states must be "NAME". If NULL, then only the state simple feature geometries are returned.}

\item{states_col}{A string that sets the column name from \code{df} containing the state names of interest. These are
full state names, either capitalized or lower case. This is a required parameter.}

\item{value_col}{A string that sets the column name from \code{df} where values(discrete or continuous) are defined.
If the column has discrete values then it must be a factor. This is a required parameter.}

\item{title}{A string that sets the plot title.}

\item{title_fontsz}{A numeric that sets the title's font size. The default is 14.}

\item{text_col}{An optional string that sets the column name from \code{df} for labelling each state polygon.}

\item{text_size}{A numeric value that sets the size of labeled state text.}

\item{text_color}{A string that sets the color of labeled state text.}

\item{text_fontface}{A string that sets the fontface of labeled state text.
Acceptable values: "plain", "bold", "italic", "bold.italic". The default is "plain".}

\item{output_dir}{A full directory path where the shapefile and its associated files will be downloaded.
The default is the directory defined by the value returned by \code{tempdir()}.}

\item{vintage}{A numeric that sets the vintage of interest. The default is 2020.}

\item{general}{A logical which if TRUE will download a less detailed, more generalized version of the state geometries.}

\item{resol}{If \code{general} is TRUE, then the resolution to return. Acceptable values are strings
"500k", "5m", "20m".}

\item{show_legend}{A logical that controls the appearance of the legend.}

\item{legend_pos}{A string that sets the legend position. Acceptable values are "right",
"top", "bottom".}

\item{na_rm}{A logical which if TRUE, missing observations are removed. If FALSE, the default,
missing observations are removed with a warning.}

\item{scale_breaks}{A string/numeric vector that defines the scale breaks. This is a required parameter.}

\item{scale_values}{A string/numeric vector that defines the possible values. For factor values, this is required
and is a vector string of colors.}

\item{scale_limits}{A string/numeric vector that defines the scale limits. This is a required parameter.}

\item{scale_labels}{An optional string vector that defines the scale labels. Vector must be the same length
as \code{scale_breaks}.}

\item{scale_colors}{Vector of colors to use for n-color gradient.}

\item{scale_na_value}{A string that sets the color for missing values.}

\item{own_scale}{A logical which if TRUE, then your own scaling may be appended to the plot without using the above
scale_* parameters.}

\item{sf_color}{A string that sets the polygon border line color.}

\item{sf_fill}{A string that sets the polygon area fill color.}

\item{sf_linewidth}{A numeric that sets the border line thickness.}

\item{sf_alpha}{A numeric that sets the alpha level attribute of the polygon fill.}

\item{display_plot}{A logical that if TRUE will display the plot. The default is TRUE.
If FALSE, then a ggplot2 object is returned}
}
\value{
A list of ggplot2 objects if \code{display_plot} is FALSE. Included in the
list is the plot of all the states ("us_states") along with the original ggplot2
geom_sf plots of the lower 48 ("lower_48"), Alaska ("alaska"), Hawaii ("hawaii") and Puerto Rico ("puerto_rico").
}
\description{
This function produces a ggplot2 based choropleth map of a discrete/continuous variable across all/selected US states
including viewable placement of Alaska, Hawaii, and Puerto Rico.  The function accepts a data frame with a column of
state names and a column with their respective values.  If the data frame (parameter \emph{df}) is submitted as NULL then only the simple
feature state geometries are returned for mapping. The function offers several options for
control/selection of state geographies and variable scaling.
}
\details{
This function depends extensively on \code{RcensusPkg::tiger_states_sf()} for obtaining state geometries, so many of
that function's parameters are repeated in this function. Also \code{RspatialPkg::get_geom_sf()} is called upon for
displaying the shapefile geometries.
}
\author{
Rick Dean
}
