% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vote2020}
\alias{vote2020}
\title{vote2020 data}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 61 rows and 22 columns.
}
\usage{
vote2020
}
\description{
Dataframe shows national presidential voting results for 2020 by state.
Letters "D" and "R" under the Party column indicate respectively Democrate or Republican
win for a state.
}
\details{
\tabular{ll}{
\strong{State} \tab character string full state name \cr
\strong{Party} \tab character string with values "D" or "R"
}
}
\keyword{datasets}
